checklist/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
connection==2021.7.20
crispy-bootstrap4==2022.1
Django==4.1.7
django-bootstrap4==22.3
django-crispy-forms==2.0
django-multiselectfield==0.1.12
idna==3.4
mysqlclient==2.1.1
numpy==1.24.2
pandas==1.5.3
pyrfc==2.8.3
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
